+++ CHANGELOG.rst
@@ -10,8 +10,9 @@ GEOPHIRES v3 (2023-2025)
 
 `release <https://github.com/NREL/GEOPHIRES-X/releases/tag/v3.9.7>`__
 
-v3.9 adds the `SAM Single Owner PPA Economic Model <https://nrel.github.io/GEOPHIRES-X/SAM-Economic-Models.html>`__
+v3.9.47 adds `Add-Ons support for SAM Economic Models <https://nrel.github.io/GEOPHIRES-X/SAM-Economic-Models.html#add-ons>`__
 
+v3.9 adds the `SAM Single Owner PPA Economic Model <https://nrel.github.io/GEOPHIRES-X/SAM-Economic-Models.html>`__
 
 3.8
 ^^^